Check to see if the log exists before rsync'ing it.

+++ admin/bin/push_to_splunk.sh
@@ -4,4 +4,7 @@ HOSTNAME=`hostname`
 SRC=/var/log/nginx/archive/musicbrainz-full-access.log-`date +%Y%m%d`
 DEST=logs@pino.mb:incoming/$HOSTNAME
 
-rsync -z $SRC $DEST
+if [ -e $SRC ]
+then
+	rsync -z $SRC $DEST
+fi
